refactor pause/resume ops into lifecycle pkg
