+++ cmd/vclusterctl/cmd/connect.go
@@ -3,13 +3,8 @@ package cmd
 import (
 	"context"
 	"fmt"
-	"github.com/loft-sh/vcluster/cmd/vclusterctl/cmd/app/localkubernetes"
-	"github.com/loft-sh/vcluster/cmd/vclusterctl/cmd/find"
-	"github.com/loft-sh/vcluster/pkg/util/translate"
 	"io"
 	"io/ioutil"
-	authenticationv1 "k8s.io/api/authentication/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 	"os"
 	"os/exec"
 	"os/signal"
@@ -19,22 +14,28 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/loft-sh/vcluster/pkg/upgrade"
-	"github.com/loft-sh/vcluster/pkg/util/portforward"
+	"github.com/pkg/errors"
+	"github.com/spf13/cobra"
+	authenticationv1 "k8s.io/api/authentication/v1"
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
-
-	"github.com/loft-sh/vcluster/cmd/vclusterctl/flags"
-	"github.com/loft-sh/vcluster/cmd/vclusterctl/log"
-	"github.com/pkg/errors"
-	"github.com/spf13/cobra"
 	"k8s.io/client-go/tools/clientcmd"
 	"k8s.io/client-go/tools/clientcmd/api"
+
+	"github.com/loft-sh/vcluster/cmd/vclusterctl/cmd/app/localkubernetes"
+	"github.com/loft-sh/vcluster/cmd/vclusterctl/cmd/find"
+	"github.com/loft-sh/vcluster/cmd/vclusterctl/flags"
+	"github.com/loft-sh/vcluster/cmd/vclusterctl/log"
+	"github.com/loft-sh/vcluster/pkg/lifecycle"
+	"github.com/loft-sh/vcluster/pkg/upgrade"
+	"github.com/loft-sh/vcluster/pkg/util/portforward"
+	"github.com/loft-sh/vcluster/pkg/util/translate"
 )
 
 // ConnectCmd holds the cmd flags
@@ -300,7 +301,7 @@ func (cmd *ConnectCmd) prepare(vclusterName string) error {
 	// resume vcluster if necessary
 	if vCluster != nil && vCluster.Status == find.StatusPaused {
 		cmd.Log.Infof("Resume vcluster %s...", vCluster.Name)
-		err = resumeVCluster(cmd.kubeClient, vclusterName, cmd.Namespace, cmd.Log)
+		err = lifecycle.ResumeVCluster(cmd.kubeClient, vclusterName, cmd.Namespace, cmd.Log)
 		if err != nil {
 			return errors.Wrap(err, "resume vcluster")
 		}
