+++ cmd/vclusterctl/cmd/resume.go
@@ -1,18 +1,15 @@
 package cmd
 
 import (
-	"context"
 	"fmt"
+
+	"github.com/spf13/cobra"
+	"k8s.io/client-go/kubernetes"
+
 	"github.com/loft-sh/vcluster/cmd/vclusterctl/cmd/find"
 	"github.com/loft-sh/vcluster/cmd/vclusterctl/flags"
 	"github.com/loft-sh/vcluster/cmd/vclusterctl/log"
-	"github.com/loft-sh/vcluster/pkg/constants"
-	"github.com/pkg/errors"
-	"github.com/spf13/cobra"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"strconv"
+	"github.com/loft-sh/vcluster/pkg/lifecycle"
 )
 
 // ResumeCmd holds the cmd flags
@@ -60,7 +57,7 @@ func (cmd *ResumeCmd) Run(args []string) error {
 		return err
 	}
 
-	err = resumeVCluster(cmd.kubeClient, args[0], cmd.Namespace, cmd.Log)
+	err = lifecycle.ResumeVCluster(cmd.kubeClient, args[0], cmd.Namespace, cmd.Log)
 	if err != nil {
 		return err
 	}
@@ -89,130 +86,4 @@ func (cmd *ResumeCmd) prepare(vClusterName string) error {
 	cmd.Namespace = vCluster.Namespace
 	cmd.kubeClient = kubeClient
 	return nil
-}
-
-func resumeVCluster(kubeClient *kubernetes.Clientset, name, namespace string, log log.Logger) error {
-	// scale down vcluster itself
-	labelSelector := "app=vcluster,release=" + name
-	found, err := scaleUpStatefulSet(kubeClient, labelSelector, namespace, log)
-	if err != nil {
-		return err
-	} else if !found {
-		found, err = scaleUpDeployment(kubeClient, labelSelector, namespace, log)
-		if err != nil {
-			return err
-		} else if !found {
-			return errors.Errorf("couldn't find a paused vcluster %s in namespace %s. Make sure the vcluster exists and was paused previously", name, namespace)
-		}
-
-		// scale down kube api server
-		_, err = scaleUpDeployment(kubeClient, "app=vcluster-api,release="+name, namespace, log)
-		if err != nil {
-			return err
-		}
-
-		// scale down kube controller
-		_, err = scaleUpDeployment(kubeClient, "app=vcluster-controller,release="+name, namespace, log)
-		if err != nil {
-			return err
-		}
-
-		// scale down etcd
-		_, err = scaleUpStatefulSet(kubeClient, "app=vcluster-etcd,release="+name, namespace, log)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func scaleUpDeployment(kubeClient kubernetes.Interface, labelSelector string, namespace string, log log.Logger) (bool, error) {
-	list, err := kubeClient.AppsV1().Deployments(namespace).List(context.TODO(), metav1.ListOptions{LabelSelector: labelSelector})
-	if err != nil {
-		return false, err
-	} else if len(list.Items) == 0 {
-		return false, nil
-	}
-
-	for _, item := range list.Items {
-		if item.Annotations == nil || item.Annotations[constants.PausedAnnotation] != "true" {
-			return false, nil
-		}
-
-		originalObject := item.DeepCopy()
-
-		replicas := 1
-		if item.Annotations[PausedReplicasAnnotation] != "" {
-			replicas, err = strconv.Atoi(item.Annotations[PausedReplicasAnnotation])
-			if err != nil {
-				log.Warnf("error parsing old replicas: %v", err)
-				replicas = 1
-			}
-		}
-
-		replicas32 := int32(replicas)
-		delete(item.Annotations, constants.PausedAnnotation)
-		delete(item.Annotations, PausedReplicasAnnotation)
-		item.Spec.Replicas = &replicas32
-
-		patch := client.MergeFrom(originalObject)
-		data, err := patch.Data(&item)
-		if err != nil {
-			return false, errors.Wrap(err, "create deployment patch")
-		}
-
-		// patch deployment
-		_, err = kubeClient.AppsV1().Deployments(namespace).Patch(context.TODO(), item.Name, patch.Type(), data, metav1.PatchOptions{})
-		if err != nil {
-			return false, errors.Wrap(err, "patch deployment")
-		}
-	}
-
-	return true, nil
-}
-
-func scaleUpStatefulSet(kubeClient kubernetes.Interface, labelSelector string, namespace string, log log.Logger) (bool, error) {
-	list, err := kubeClient.AppsV1().StatefulSets(namespace).List(context.TODO(), metav1.ListOptions{LabelSelector: labelSelector})
-	if err != nil {
-		return false, err
-	} else if len(list.Items) == 0 {
-		return false, nil
-	}
-
-	for _, item := range list.Items {
-		if item.Annotations == nil || item.Annotations[constants.PausedAnnotation] != "true" {
-			return false, nil
-		}
-
-		originalObject := item.DeepCopy()
-
-		replicas := 1
-		if item.Annotations[PausedReplicasAnnotation] != "" {
-			replicas, err = strconv.Atoi(item.Annotations[PausedReplicasAnnotation])
-			if err != nil {
-				log.Warnf("error parsing old replicas: %v", err)
-				replicas = 1
-			}
-		}
-
-		replicas32 := int32(replicas)
-		delete(item.Annotations, constants.PausedAnnotation)
-		delete(item.Annotations, PausedReplicasAnnotation)
-		item.Spec.Replicas = &replicas32
-
-		patch := client.MergeFrom(originalObject)
-		data, err := patch.Data(&item)
-		if err != nil {
-			return false, errors.Wrap(err, "create statefulSet patch")
-		}
-
-		// patch deployment
-		_, err = kubeClient.AppsV1().StatefulSets(namespace).Patch(context.TODO(), item.Name, patch.Type(), data, metav1.PatchOptions{})
-		if err != nil {
-			return false, errors.Wrap(err, "patch statefulSet")
-		}
-	}
-
-	return true, nil
 }
