+++ pkg/constants/annotation.go
@@ -4,5 +4,6 @@ const (
 	SkipTranslationAnnotation = "vcluster.loft.sh/skip-translate"
 	SyncResourceAnnotation    = "vcluster.loft.sh/force-sync"
 
-	PausedAnnotation = "loft.sh/paused"
+	PausedAnnotation         = "loft.sh/paused"
+	PausedReplicasAnnotation = "loft.sh/paused-replicas"
 )
